requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
cx-Oracle==8.3.0
flake8==4.0.1
psycopg2-binary==2.9.2
pytest==7.1.1
pytest-env==0.6.2
pytest-rerunfailures==10.2
pytest-xdist==2.4.0
requests==2.26.0
selenium==4.1.3
SQLAlchemy==1.4.29
testrail-api==1.10.3
webdriver-manager==3.5.4
urllib3~=1.26.9
